+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/InputTableMetadata.java
@@ -1,6 +1,5 @@
 package org.dbsp.sqlCompiler.compiler;
 
-import org.dbsp.sqlCompiler.compiler.frontend.statements.CreateTableStatement;
 import org.dbsp.util.Utilities;
 
 import java.util.Collection;
@@ -9,14 +8,10 @@ import java.util.List;
 
 /** Metadata describing an input table. */
 public class InputTableMetadata {
-    public final String tableName;
-    public final CreateTableStatement statement;
     final LinkedHashMap<String, InputColumnMetadata> columnMetadata;
 
-    public InputTableMetadata(String tableName, CreateTableStatement statement, List<InputColumnMetadata> columns) {
+    public InputTableMetadata(List<InputColumnMetadata> columns) {
         this.columnMetadata = new LinkedHashMap<>();
-        this.tableName = tableName;
-        this.statement = statement;
         for (InputColumnMetadata meta: columns) {
             Utilities.putNew(this.columnMetadata, meta.name, meta);
         }