+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/ProgramMetadata.java
@@ -3,6 +3,7 @@ package org.dbsp.sqlCompiler.compiler;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import org.dbsp.sqlCompiler.compiler.frontend.statements.IHasSchema;
 import org.dbsp.util.Utilities;
 
 import java.util.LinkedHashMap;
@@ -10,8 +11,8 @@ import java.util.LinkedHashMap;
 /** Represents metadata about the compiled program.
  * Contains a description of all input tables and all views. */
 public class ProgramMetadata {
-    final LinkedHashMap<String, InputTableDescription> inputTables;
-    final LinkedHashMap<String, OutputViewDescription> outputViews;
+    final LinkedHashMap<String, IHasSchema> inputTables;
+    final LinkedHashMap<String, IHasSchema> outputViews;
 
     public ProgramMetadata() {
         this.inputTables = new LinkedHashMap<>();
@@ -21,10 +22,10 @@ public class ProgramMetadata {
     public ObjectNode asJson() {
         ObjectMapper mapper = new ObjectMapper();
         ArrayNode inputs = mapper.createArrayNode();
-        for (InputTableDescription input: this.inputTables.values())
+        for (IHasSchema input: this.inputTables.values())
             inputs.add(input.asJson());
         ArrayNode outputs = mapper.createArrayNode();
-        for (OutputViewDescription output: this.outputViews.values())
+        for (IHasSchema output: this.outputViews.values())
             outputs.add(output.asJson());
         ObjectNode ios = mapper.createObjectNode();
         ios.set("inputs", inputs);
@@ -32,19 +33,19 @@ public class ProgramMetadata {
         return ios;
     }
 
-    public InputTableDescription getTableDescription(String name) {
+    public IHasSchema getTableDescription(String name) {
         return Utilities.getExists(this.inputTables, name);
     }
 
-    public OutputViewDescription getViewDescription(String name) {
+    public IHasSchema getViewDescription(String name) {
         return Utilities.getExists(this.outputViews, name);
     }
 
-    public void addTable(InputTableDescription description) {
+    public void addTable(IHasSchema description) {
         this.inputTables.put(description.getName(), description);
     }
 
-    public void addView(OutputViewDescription description) {
+    public void addView(IHasSchema description) {
         this.outputViews.put(description.getName(), description);
     }
 }