+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/suites/NexmarkTest.java
@@ -613,6 +613,7 @@ SELECT
         }
 
         Assert.assertFalse(compiler.hasErrors());
+        Assert.assertFalse(compiler.hasWarnings());
         this.addRustTestCase("NexmarkTest", compiler, getCircuit(compiler));
     }
 }